AuthContext 수정 -AuthContextProvider warpping -AuthContext kebap case 네이밍 적용, provider 파일과 통합 -로그인 완료 후 홈으로 리다이렉트
