+++ src/app/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react'
 import '@/assets/styles/globals.scss'
 import Main from './main/page'
-import AuthContextProvider from '../store/AuthContextProvider'
+import AuthContextProvider from '../store/auth-context'
 import { Metadata } from 'next'
 
 //use the Metadata API which automatically handles advanced requirements such as streaming and de-duplicating <head> elements.
