+++ src/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import { AuthContext } from '@/store/AuthContext'
+import { AuthContext } from '@/store/auth-context'
 
 export const useAuth = () => {
   return useContext(AuthContext)
